requirements.txt
claude-agent-sdk
openai-agents
python-dotenv